@@ -94,7 +94,7 @@
     "@vercel/functions": "^2.0.0",
     "@vercel/otel": "^1.12.0",
     "@vercel/postgres": "^0.10.0",
-    "ai": "4.3.13",
+    "ai": "^4.3.16",
     "axios": "^1.10.0",
     "bcrypt-ts": "^5.0.2",
     "class-variance-authority": "^0.7.0",
@@ -115,7 +115,7 @@
     "lucide-react": "^0.446.0",
     "marked": "^16.0.0",
     "nanoid": "^5.0.8",
-    "next": "15.3.0-canary.31",
+    "next": "15.3.1",
     "next-auth": "5.0.0-beta.25",
     "next-themes": "^0.3.0",
     "orderedmap": "^2.1.1",
@@ -129,9 +129,9 @@
     "prosemirror-schema-list": "^1.4.1",
     "prosemirror-state": "^1.4.3",
     "prosemirror-view": "^1.34.3",
-    "react": "19.0.0-rc-45804af1-20241021",
+    "react": "^19.1.0",
     "react-data-grid": "7.0.0-beta.47",
-    "react-dom": "19.0.0-rc-45804af1-20241021",
+    "react-dom": "^19.1.0",
     "react-markdown": "^9.0.1",
     "react-resizable-panels": "^2.1.7",
     "redis": "^5.0.0",
@@ -144,7 +144,7 @@
     "swr": "^2.2.5",
     "tailwind-merge": "^2.5.2",
     "usehooks-ts": "^3.1.0",
-    "zod": "^3.23.8"
+    "zod": "^3.25.67"
   },
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
@@ -165,7 +165,7 @@
     "cross-env": "^7.0.3",
     "dotenv": "^17.2.1",
     "drizzle-kit": "^0.25.0",
-    "eslint": "^8.57.0",
+    "eslint": "^9.24.0",
     "eslint-config-next": "14.2.5",
     "eslint-config-prettier": "^9.1.0",
     "eslint-import-resolver-typescript": "^3.6.3",
