@@ -110,7 +110,7 @@
     "process": "^0.11.10",
     "rimraf": "^6.0.1",
     "ts-node": "^10.9.2",
-    "typescript": "^5.2.2",
+    "typescript": "^5.6.3",
     "typescript-eslint": "^8.31.1",
     "vitest": "^1.3.1",
     "wrangler": "^4.14.1"
