@@ -1,6 +1,6 @@
 {
   "name": "@turbo-super/core",
-  "version": "0.0.1",
+  "version": "1.0.0",
   "private": true,
   "main": "./dist/index.js",
   "module": "./dist/index.mjs",
@@ -14,17 +14,17 @@
   "scripts": {
     "build": "tsup",
     "dev": "tsup --watch",
-    "clean": "rm -rf dist"
+    "clean": "node ../../scripts/clean.js dist"
   },
   "dependencies": {
-    "zod": "^3.22.4"
+    "zod": "^3.25.67"
   },
   "devDependencies": {
     "@types/node": "^20.0.0",
-    "tsup": "^8.0.1",
-    "typescript": "^5.3.3"
+    "tsup": "^8.0.2",
+    "typescript": "^5.6.3"
   },
   "peerDependencies": {
-    "next": "^15.0.0"
+    "next": "15.3.1"
   }
 }
