@@ -1,6 +1,6 @@
 {
   "name": "@turbo-super/features",
-  "version": "0.0.1",
+  "version": "1.0.0",
   "private": true,
   "main": "./dist/index.js",
   "module": "./dist/index.mjs",
@@ -26,7 +26,7 @@
   "scripts": {
     "build": "tsup",
     "dev": "tsup --watch",
-    "clean": "rm -rf dist"
+    "clean": "node ../../scripts/clean.js dist"
   },
   "dependencies": {
     "@radix-ui/colors": "^1.0.1",
@@ -46,11 +46,11 @@
     "lodash-es": "^4.17.21",
     "lucide-react": "^0.400.0",
     "minimatch": "^10.0.3",
-    "react": "^18.0.0 || ^19.0.0",
-    "react-dom": "^18.0.0 || ^19.0.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "remotion": "4.0.229",
     "super-timeline": "https://github.com/PranovAdilet/super-timeline-vite2.git#6992427",
-    "zod": "^3.22.4",
+    "zod": "^3.25.67",
     "zustand": "^4.5.4"
   },
   "devDependencies": {
@@ -59,11 +59,11 @@
     "@types/react": "^18.0.0 || ^19.0.0",
     "@types/react-dom": "^18.0.0 || ^19.0.0",
     "@types/tailwindcss": "^3.1.0",
-    "tsup": "^8.0.1",
-    "typescript": "^5.3.3"
+    "tsup": "^8.0.2",
+    "typescript": "^5.6.3"
   },
   "peerDependencies": {
-    "next": "^15.0.0",
+    "next": "15.3.1",
     "tailwindcss": "^3.4.0"
   }
 }
