@@ -1,7 +1,8 @@
 {
   "name": "@turbo-super/ui",
-  "version": "0.0.0",
+  "version": "1.0.0",
   "private": true,
+  "type": "module",
   "description": "Общие UI компоненты для turbo-super",
   "main": "./dist/index.js",
   "module": "./dist/index.mjs",
@@ -19,17 +20,17 @@
     "dev": "tsup --watch",
     "type-check": "tsc --noEmit",
     "lint": "eslint src/",
-    "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist"
+    "clean": "node ../../scripts/clean.js .turbo node_modules dist"
   },
   "peerDependencies": {
-    "react": "^18.0.0 || ^19.0.0",
-    "react-dom": "^18.0.0 || ^19.0.0"
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0"
   },
   "devDependencies": {
     "@turbo-super/eslint-config": "workspace:*",
     "@types/react": "^18.2.61",
     "@types/react-dom": "^18.2.19",
-    "eslint": "^8.57.0",
+    "eslint": "^9.24.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "tsup": "^8.0.2",
