@@ -36,16 +36,17 @@ export default [
       'node_modules/**',
       'public/**',
       '**/*.config.js',
+      '**/*.config.mjs',
       'next-env.d.ts',
       '.open-next/**',
       'out/**',
       '.contentlayer/**',
       '.wrangler/**',
-      'scripts/optimize-veo3-image.js'
+      'scripts/**'
     ]
   },
   {
-    files: ['**/*.{js,mjs,jsx,ts,tsx}'],
+    files: ['**/*.{js,mjs,jsx}'],
     languageOptions: {
       ecmaVersion: 2022,
       sourceType: 'module',
@@ -55,7 +56,16 @@ export default [
         }
       },
       globals: {
-        process: 'readonly'
+        process: 'readonly',
+        console: 'readonly',
+        require: 'readonly',
+        __dirname: 'readonly',
+        __filename: 'readonly',
+        module: 'readonly',
+        exports: 'readonly',
+        global: 'readonly',
+        Buffer: 'readonly',
+        fetch: 'readonly'
       }
     },
     // Настройки для React дублируем здесь для надежности
@@ -72,11 +82,114 @@ export default [
       '@next/next/no-page-custom-font': 'off',
       'react/react-in-jsx-scope': 'off',
       'react/prop-types': 'off',
-      // Разрешаем неиспользуемые переменные в определенных случаях
+      
+      // Строгие правила TypeScript (только для файлов с type information)
+      '@typescript-eslint/no-explicit-any': 'warn',
       '@typescript-eslint/no-unused-vars': ['error', { 
         'argsIgnorePattern': '^_',
         'varsIgnorePattern': '^_', 
         'ignoreRestSiblings': true 
+      }],
+      '@typescript-eslint/prefer-nullish-coalescing': 'warn',
+      '@typescript-eslint/prefer-optional-chain': 'warn',
+      '@typescript-eslint/no-unnecessary-type-assertion': 'warn',
+      '@typescript-eslint/no-non-null-assertion': 'warn',
+      
+      // React Hooks правила
+      'react-hooks/rules-of-hooks': 'error',
+      'react-hooks/exhaustive-deps': 'error',
+      
+      // Общие правила
+      'no-console': 'warn',
+      'no-debugger': 'error',
+      'no-alert': 'warn',
+      'no-var': 'error',
+      'prefer-const': 'error',
+      'no-unused-expressions': 'error',
+      'no-duplicate-imports': 'warn',
+      'no-useless-return': 'error',
+      'no-useless-constructor': 'error',
+      'no-useless-catch': 'error',
+      'no-useless-concat': 'error',
+      'no-useless-escape': 'error',
+      'no-useless-rename': 'error',
+      'no-void': 'error',
+      'prefer-template': 'warn',
+      'prefer-arrow-callback': 'error',
+      'prefer-destructuring': ['warn', {
+        'array': true,
+        'object': true
+      }, {
+        'enforceForRenamedProperties': false
+      }]
+    }
+  },
+  {
+    files: ['**/*.{ts,tsx}'],
+    languageOptions: {
+      ecmaVersion: 2022,
+      sourceType: 'module',
+      parserOptions: {
+        ecmaFeatures: {
+          jsx: true
+        },
+        project: './tsconfig.json',
+        tsconfigRootDir: __dirname
+      },
+      globals: {
+        process: 'readonly',
+        console: 'readonly',
+        fetch: 'readonly'
+      }
+    },
+    settings: {
+      react: {
+        version: '19.0'
+      }
+    },
+    rules: {
+      // Строгие правила TypeScript для файлов с type information
+      '@typescript-eslint/no-explicit-any': 'warn',
+      '@typescript-eslint/no-unused-vars': ['error', { 
+        'argsIgnorePattern': '^_',
+        'varsIgnorePattern': '^_', 
+        'ignoreRestSiblings': true 
+      }],
+      '@typescript-eslint/no-unsafe-assignment': 'warn',
+      '@typescript-eslint/no-unsafe-member-access': 'warn',
+      '@typescript-eslint/no-unsafe-call': 'warn',
+      '@typescript-eslint/no-unsafe-return': 'warn',
+      '@typescript-eslint/prefer-nullish-coalescing': 'warn',
+      '@typescript-eslint/prefer-optional-chain': 'warn',
+      '@typescript-eslint/no-unnecessary-type-assertion': 'warn',
+      '@typescript-eslint/no-non-null-assertion': 'warn',
+      
+      // React Hooks правила
+      'react-hooks/rules-of-hooks': 'error',
+      'react-hooks/exhaustive-deps': 'error',
+      
+      // Общие правила
+      'no-console': 'warn',
+      'no-debugger': 'error',
+      'no-alert': 'warn',
+      'no-var': 'error',
+      'prefer-const': 'error',
+      'no-unused-expressions': 'error',
+      'no-duplicate-imports': 'warn',
+      'no-useless-return': 'error',
+      'no-useless-constructor': 'error',
+      'no-useless-catch': 'error',
+      'no-useless-concat': 'error',
+      'no-useless-escape': 'error',
+      'no-useless-rename': 'error',
+      'no-void': 'error',
+      'prefer-template': 'warn',
+      'prefer-arrow-callback': 'error',
+      'prefer-destructuring': ['warn', {
+        'array': true,
+        'object': true
+      }, {
+        'enforceForRenamedProperties': false
       }]
     }
   }
