@@ -1,12 +1,48 @@
 import js from "@eslint/js";
+import typescript from '@typescript-eslint/eslint-plugin';
+import typescriptParser from '@typescript-eslint/parser';
+import prettier from 'eslint-config-prettier';
 
 export default [
     {
         ignores: [
             "src/superduperai/api/**",
+            "src/superduperai/lib/**",
             "dist/**",
-            "node_modules/**"
+            "node_modules/**",
+            "scripts/**"
         ]
     },
-    js.configs.recommended
+    js.configs.recommended,
+    {
+        files: ['**/*.{js,ts,tsx}'],
+        languageOptions: {
+            parser: typescriptParser,
+            parserOptions: {
+                ecmaVersion: 2022,
+                sourceType: 'module'
+            },
+            globals: {
+                console: 'readonly',
+                process: 'readonly',
+                Buffer: 'readonly',
+                __dirname: 'readonly',
+                __filename: 'readonly',
+                global: 'readonly',
+                module: 'readonly',
+                require: 'readonly',
+                exports: 'readonly'
+            }
+        },
+        plugins: {
+            '@typescript-eslint': typescript
+        },
+        rules: {
+            ...typescript.configs.recommended.rules,
+            ...prettier.rules,
+            '@typescript-eslint/no-unused-vars': 'error',
+            '@typescript-eslint/no-explicit-any': 'warn',
+            'prefer-const': 'error'
+        }
+    }
 ];
